Extract workflow and event update logic to Clients
