+++ lib/focal_api/clients.ex
@@ -4,6 +4,8 @@ defmodule FocalApi.Clients do
   alias FocalApi.Clients.Client
   alias FocalApi.Accounts
   alias FocalApi.Clients.Event
+  alias FocalApi.EventName
+  alias FocalApi.DefaultWorkflows
 
   def list_clients do
     Repo.all(Client)
@@ -242,5 +244,60 @@ defmodule FocalApi.Clients do
 
   def change_workflow(%Workflow{} = workflow) do
     Workflow.changeset(workflow, %{})
+  end
+
+  def handle_engagement_and_wedding_updates(package) do
+    client = get_client!(package.client_id)
+    events = list_events_by_package(package.uuid)
+    workflows = list_workflows_by_client(client.uuid)
+    workflow_has_engagement = workflows_include(workflows, EventName.engagement())
+    workflow_has_wedding = workflows_include(workflows, EventName.wedding())
+
+    if include_engagement?(package, workflow_has_engagement), do: DefaultWorkflows.create_engagement_workflow_and_tasks(client, 2)
+    if include_wedding?(package, workflow_has_wedding), do: DefaultWorkflows.create_wedding_workflow_and_tasks(client, 3)
+    if remove_engagement?(package, workflow_has_engagement), do: delete_workflow_and_event(workflows, events, EventName.engagement())
+    if remove_wedding?(package, workflow_has_wedding), do: delete_workflow_and_event(workflows, events, EventName.wedding())
+  end
+
+  defp workflows_include(workflows, workflow_name) do
+    workflows
+    |> Enum.any?(fn workflow -> workflow.workflow_name == workflow_name end)
+  end
+
+  defp include_engagement?(package, workflow_has_engagement) do
+    package.engagement_included == true && !workflow_has_engagement
+  end
+
+  defp include_wedding?(package, workflow_has_wedding) do
+    package.wedding_included == true && !workflow_has_wedding
+  end
+
+  defp remove_engagement?(package, workflow_has_engagement) do
+    package.engagement_included == false && workflow_has_engagement
+  end
+
+  defp remove_wedding?(package, workflow_has_wedding) do
+    package.wedding_included == false && workflow_has_wedding
+  end
+
+  defp delete_workflow_and_event(workflows, events, name) do
+    delete_workflow_by_name(workflows, name)
+    delete_event_by_name(events, name)
+  end
+
+  defp delete_workflow_by_name(workflows, workflow_name) do
+    workflows
+    |> Enum.find(fn workflow -> workflow.workflow_name == workflow_name end)
+    |> delete_workflow()
+  end
+
+  defp delete_event_by_name(events, event_name) do
+    event = events
+    |> Enum.find(fn event -> event.event_name == event_name end)
+
+    if (event != nil) do
+      event
+      |> delete_event()
+    end
   end
 end
