+++ lib/focal_api_web/controllers/package_controller.ex
@@ -4,7 +4,6 @@ defmodule FocalApiWeb.PackageController do
   alias FocalApi.Clients
   alias FocalApi.Clients.Package
   alias FocalApi.DefaultWorkflows
-  alias FocalApi.EventName
 
   action_fallback FocalApiWeb.FallbackController
 
@@ -48,7 +47,7 @@ defmodule FocalApiWeb.PackageController do
     |> Map.put("uuid", package.uuid)
 
     with {:ok, %Package{} = package} <- Clients.update_package(package, update_attrs) do
-      handle_engagement_and_wedding_updates(package)
+      Clients.handle_engagement_and_wedding_updates(package)
       render(conn, "show.json", package: package)
     end
   end
@@ -58,61 +57,6 @@ defmodule FocalApiWeb.PackageController do
 
     with {:ok, %Package{}} <- Clients.delete_package(package) do
       send_resp(conn, :no_content, "")
-    end
-  end
-
-  defp handle_engagement_and_wedding_updates(package) do
-    client = Clients.get_client!(package.client_id)
-    events = Clients.list_events_by_package(package.uuid)
-    workflows = Clients.list_workflows_by_client(client.uuid)
-    workflow_has_engagement = workflows_include(workflows, EventName.engagement())
-    workflow_has_wedding = workflows_include(workflows, EventName.wedding())
-
-    if include_engagement?(package, workflow_has_engagement), do: DefaultWorkflows.create_engagement_workflow_and_tasks(client, 2)
-    if include_wedding?(package, workflow_has_wedding), do: DefaultWorkflows.create_wedding_workflow_and_tasks(client, 3)
-    if remove_engagement?(package, workflow_has_engagement), do: delete_workflow_and_event(workflows, events, EventName.engagement())
-    if remove_wedding?(package, workflow_has_wedding), do: delete_workflow_and_event(workflows, events, EventName.wedding())
-  end
-
-  defp workflows_include(workflows, workflow_name) do
-    workflows
-    |> Enum.any?(fn workflow -> workflow.workflow_name == workflow_name end)
-  end
-
-  defp include_engagement?(package, workflow_has_engagement) do
-    package.engagement_included == true && !workflow_has_engagement
-  end
-
-  defp include_wedding?(package, workflow_has_wedding) do
-    package.wedding_included == true && !workflow_has_wedding
-  end
-
-  defp remove_engagement?(package, workflow_has_engagement) do
-    package.engagement_included == false && workflow_has_engagement
-  end
-
-  defp remove_wedding?(package, workflow_has_wedding) do
-    package.wedding_included == false && workflow_has_wedding
-  end
-
-  defp delete_workflow_and_event(workflows, events, name) do
-    delete_workflow(workflows, name)
-    delete_event(events, name)
-  end
-
-  defp delete_workflow(workflows, workflow_name) do
-    workflows
-    |> Enum.find(fn workflow -> workflow.workflow_name == workflow_name end)
-    |> Clients.delete_workflow()
-  end
-
-  defp delete_event(events, event_name) do
-    event = events
-    |> Enum.find(fn event -> event.event_name == event_name end)
-
-    if (event != nil) do
-      event
-      |> Clients.delete_event()
     end
   end
 end
